Move webpack config to each app
