+++ apps/drupal/webpack.drupal.config.js
@@ -9,13 +9,12 @@ const merge = require('webpack-merge');
 const webpack = require('webpack');
 
 // Custom Imports
-const shared = require('./webpack.shared.config');
-const { PATH_DRUPAL } = require('./config');
+const shared = require('../../webpack.shared.config');
 
 const drupal = {
   entry: {
     'app-drupal': [
-      path.resolve(__dirname, PATH_DRUPAL, 'index.js'),
+      path.resolve(__dirname, 'index.js'),
     ],
   },
   // These will be explicitly be provided OUTSIDE the bundle via a <script> tag in the HTML
